+++ spec/mongoid/relations/metadata_spec.rb
@@ -442,8 +442,6 @@ describe Mongoid::Relations::Metadata do
             metadata.class_name.should eq("Address")
           end
         end
-
-
       end
     end
   end
@@ -1416,7 +1414,21 @@ describe Mongoid::Relations::Metadata do
     end
   end
 
-  context "#order" do
+  describe "#options" do
+
+    let(:metadata) do
+      described_class.new(
+        :order => :rating.asc,
+        :relation => Mongoid::Relations::Referenced::Many
+      )
+    end
+
+    it "returns self" do
+      metadata.options.should eq(metadata)
+    end
+  end
+
+  describe "#order" do
 
     let(:metadata) do
       described_class.new(
@@ -1428,7 +1440,6 @@ describe Mongoid::Relations::Metadata do
     it "returns order criteria" do
       metadata.order.should eq(:rating.asc)
     end
-
   end
 
   describe "#klass" do